+++ docs/pages/pages.11tydata.js
@@ -1,0 +1,6 @@
+module.exports = {
+  layout: 'page',
+  eleventyComputed: {
+    permalink: (article) => `/${article.page.fileSlug}/index.html`,
+  },
+};