+++ docs/site/_data/env.js
@@ -1,0 +1,3 @@
+module.exports = {
+  dev: process.env.NODE_ENV === 'development',
+};