+++ docs/src/js/component-sidebar.js
@@ -1,0 +1,5 @@
+document.querySelector('.component-sidebar-menu-button').onclick = () => {
+  document
+    .querySelector('.component-sidebar-menu-content')
+    .classList.toggle('component-sidebar-menu-content-reveal');
+};