+++ docs/src/js/index.js
@@ -1,0 +1,6 @@
+/* eslint import/no-unresolved: 'off' */
+import '../../../components/menu/src/index.js';
+import '../../../components/content-navigation/src/index.js';
+import '../../../components/plus/index.js';
+import '../../../components/minus/index.js';
+import './component-sidebar.js';