+++ components/back-to-top/CHANGELOG.md
@@ -1,4 +1,5 @@
-# CHANGELOG for ds-back-to-top
+# Navigation components >> Back to top button >> Changelog || 30
+
 `ds-back-to-to`
 
 # 1.0.2
@@ -6,7 +7,7 @@
 * Added unit test.
 
 # 1.0.1
-* Changed button's animating in and out of the side of the screen insted of bottom.
+* Changed button's animating in and out of the side of the screen instead of bottom.
 
 # 1.0.0
 * Initial commit